Moves cmd/*.go to sub folder with cli name - as it's best practice - now /cmd/reliably/*.go
