+++ cmd/reliably/auth/logout.go
