+++ cmd/reliably/cmdutil/auth_check.go
