+++ cmd/reliably/completion.go
