+++ cmd/reliably/doc/man.go
